+++ src/command_modules/azure-cli-keyvault/azure/cli/command_modules/keyvault/__init__.py
@@ -12,8 +12,8 @@ class KeyVaultCommandsLoader(AzCommandsLoader):
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
-        from azure.cli.command_modules.keyvault._client_factory import keyvault_client_factory
-        from azure.cli.command_modules.keyvault._command_type import KeyVaultCommandGroup, KeyVaultArgumentContext
+        from ._client_factory import keyvault_client_factory
+        from ._command_type import KeyVaultCommandGroup, KeyVaultArgumentContext
         keyvault_custom = CliCommandType(
             operations_tmpl='azure.cli.command_modules.keyvault.custom#{}',
             client_factory=keyvault_client_factory
@@ -24,12 +24,12 @@ class KeyVaultCommandsLoader(AzCommandsLoader):
                                                      argument_context_cls=KeyVaultArgumentContext)
 
     def load_command_table(self, args):
-        from azure.cli.command_modules.keyvault.commands import load_command_table
+        from .commands import load_command_table
         load_command_table(self, args)
         return self.command_table
 
     def load_arguments(self, command):
-        from azure.cli.command_modules.keyvault._params import load_arguments
+        from ._params import load_arguments
         load_arguments(self, command)